[Checks] split main from types
